requirements.txt
# Core requirements
lxml>=4.9.0  # For XML parsing 
python-dotenv>=1.0.0  # For environment variable management

# Database requirements
psycopg2-binary>=2.9.9  # PostgreSQL adapter
SQLAlchemy>=2.0.0  # SQL toolkit and ORM
alembic>=1.13.0  # Database migration tool 